Add new conjoined SendTransactionService
