+++ Cargo.toml
@@ -63,6 +63,7 @@ members = [
     "stake-accounts",
     "sys-tuner",
     "tokens",
+    "transaction",
     "transaction-status",
     "account-decoder",
     "upload-perf",
