+++ banks-server/Cargo.toml
@@ -17,7 +17,7 @@ mio = "0.7.13"
 solana-banks-interface = { path = "../banks-interface", version = "=1.8.0" }
 solana-runtime = { path = "../runtime", version = "=1.8.0" }
 solana-sdk = { path = "../sdk", version = "=1.8.0" }
-solana-metrics = { path = "../metrics", version = "=1.8.0" }
+solana-transaction = { path = "../transaction", version = "=1.8.0" }
 tarpc = { version = "0.26.2", features = ["full"] }
 tokio = { version = "1", features = ["full"] }
 tokio-serde = { version = "0.8", features = ["bincode"] }
