+++ banks-server/src/lib.rs
@@ -1,7 +1,3 @@
 #![allow(clippy::integer_arithmetic)]
 pub mod banks_server;
 pub mod rpc_banks_service;
-pub mod send_transaction_service;
-
-#[macro_use]
-extern crate solana_metrics;
