+++ rpc/Cargo.toml
@@ -41,6 +41,7 @@ solana-runtime = { path = "../runtime", version = "=1.8.0" }
 solana-sdk = { path = "../sdk", version = "=1.8.0" }
 solana-streamer = { path = "../streamer", version = "=1.8.0" }
 solana-storage-bigtable = { path = "../storage-bigtable", version = "=1.8.0" }
+solana-transaction = { path = "../transaction", version = "=1.8.0" }
 solana-transaction-status = { path = "../transaction-status", version = "=1.8.0" }
 solana-version = { path = "../version", version = "=1.8.0" }
 solana-vote-program = { path = "../programs/vote", version = "=1.8.0" }
