+++ transaction/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "solana-transaction"
+version = "1.8.0"
+description = "Solana transaction"
+authors = ["Solana Maintainers <maintainers@solana.foundation>"]
+repository = "https://github.com/solana-labs/solana"
+homepage = "https://solana.com/"
+documentation = "https://docs.rs/solana-transaction-status"
+license = "Apache-2.0"
+edition = "2018"
+
+[dependencies]
+log = "0.4.14"
+solana-logger = { path = "../logger", version = "=1.8.0" }
+solana-metrics = { path = "../metrics", version = "=1.8.0" }
+solana-runtime = { path = "../runtime", version = "=1.8.0" }
+solana-sdk = { path = "../sdk", version = "=1.8.0" }
+
+[package.metadata.docs.rs]
+targets = ["x86_64-unknown-linux-gnu"]
