+++ transaction/src/lib.rs
@@ -1,0 +1,6 @@
+#![allow(clippy::integer_arithmetic)]
+pub mod send_transaction_service;
+pub mod tpu_info;
+
+#[macro_use]
+extern crate solana_metrics;
