+++ transaction/src/send_transaction_service.rs
@@ -1,20 +1,16 @@
 // TODO: Merge this implementation with the one at `banks-server/src/send_transaction_service.rs`
 use {
+    crate::tpu_info::TpuInfo,
     log::*,
-    solana_gossip::cluster_info::ClusterInfo,
     solana_metrics::{datapoint_warn, inc_new_counter_info},
-    solana_poh::poh_recorder::PohRecorder,
     solana_runtime::{bank::Bank, bank_forks::BankForks},
-    solana_sdk::{
-        clock::NUM_CONSECUTIVE_LEADER_SLOTS, hash::Hash, nonce_account, pubkey::Pubkey,
-        signature::Signature,
-    },
+    solana_sdk::{hash::Hash, nonce_account, pubkey::Pubkey, signature::Signature},
     std::{
         collections::HashMap,
         net::{SocketAddr, UdpSocket},
         sync::{
             mpsc::{Receiver, RecvTimeoutError},
-            Arc, Mutex, RwLock,
+            Arc, RwLock,
         },
         thread::{self, Builder, JoinHandle},
         time::{Duration, Instant},
@@ -51,48 +47,6 @@ impl TransactionInfo {
     }
 }
 
-pub struct LeaderInfo {
-    cluster_info: Arc<ClusterInfo>,
-    poh_recorder: Arc<Mutex<PohRecorder>>,
-    recent_peers: HashMap<Pubkey, SocketAddr>,
-}
-
-impl LeaderInfo {
-    pub fn new(cluster_info: Arc<ClusterInfo>, poh_recorder: Arc<Mutex<PohRecorder>>) -> Self {
-        Self {
-            cluster_info,
-            poh_recorder,
-            recent_peers: HashMap::new(),
-        }
-    }
-
-    pub fn refresh_recent_peers(&mut self) {
-        self.recent_peers = self
-            .cluster_info
-            .tpu_peers()
-            .into_iter()
-            .map(|ci| (ci.id, ci.tpu))
-            .collect();
-    }
-
-    pub fn get_leader_tpus(&self, max_count: u64) -> Vec<&SocketAddr> {
-        let recorder = self.poh_recorder.lock().unwrap();
-        let leaders: Vec<_> = (0..max_count)
-            .filter_map(|i| recorder.leader_after_n_slots(i * NUM_CONSECUTIVE_LEADER_SLOTS))
-            .collect();
-        drop(recorder);
-        let mut unique_leaders = vec![];
-        for leader in leaders.iter() {
-            if let Some(addr) = self.recent_peers.get(leader) {
-                if !unique_leaders.contains(&addr) {
-                    unique_leaders.push(addr);
-                }
-            }
-        }
-        unique_leaders
-    }
-}
-
 #[derive(Default, Debug, PartialEq)]
 struct ProcessTransactionsResult {
     rooted: u64,
@@ -103,10 +57,10 @@ struct ProcessTransactionsResult {
 }
 
 impl SendTransactionService {
-    pub fn new(
+    pub fn new<T: TpuInfo + std::marker::Send + 'static>(
         tpu_address: SocketAddr,
         bank_forks: &Arc<RwLock<BankForks>>,
-        leader_info: Option<LeaderInfo>,
+        leader_info: Option<T>,
         receiver: Receiver<TransactionInfo>,
         retry_rate_ms: u64,
         leader_forward_count: u64,
@@ -122,11 +76,11 @@ impl SendTransactionService {
         Self { thread }
     }
 
-    fn retry_thread(
+    fn retry_thread<T: TpuInfo + std::marker::Send + 'static>(
         tpu_address: SocketAddr,
         receiver: Receiver<TransactionInfo>,
         bank_forks: Arc<RwLock<BankForks>>,
-        mut leader_info: Option<LeaderInfo>,
+        mut leader_info: Option<T>,
         retry_rate_ms: u64,
         leader_forward_count: u64,
     ) -> JoinHandle<()> {
@@ -209,13 +163,13 @@ impl SendTransactionService {
             .unwrap()
     }
 
-    fn process_transactions(
+    fn process_transactions<T: TpuInfo>(
         working_bank: &Arc<Bank>,
         root_bank: &Arc<Bank>,
         send_socket: &UdpSocket,
         tpu_address: &SocketAddr,
         transactions: &mut HashMap<Signature, TransactionInfo>,
-        leader_info: &Option<LeaderInfo>,
+        leader_info: &Option<T>,
         leader_forward_count: u64,
     ) -> ProcessTransactionsResult {
         let mut result = ProcessTransactionsResult::default();
@@ -312,26 +266,13 @@ impl SendTransactionService {
 mod test {
     use {
         super::*,
-        solana_gossip::contact_info::ContactInfo,
-        solana_ledger::{
-            blockstore::Blockstore, get_tmp_ledger_path, leader_schedule_cache::LeaderScheduleCache,
-        },
-        solana_runtime::genesis_utils::{
-            create_genesis_config_with_vote_accounts, GenesisConfigInfo, ValidatorVoteKeypairs,
-        },
+        crate::tpu_info::NullTpuInfo,
         solana_sdk::{
-            account::AccountSharedData,
-            fee_calculator::FeeCalculator,
-            genesis_config::create_genesis_config,
-            nonce,
-            poh_config::PohConfig,
-            pubkey::Pubkey,
-            signature::{Keypair, Signer},
+            account::AccountSharedData, fee_calculator::FeeCalculator,
+            genesis_config::create_genesis_config, nonce, pubkey::Pubkey, signature::Signer,
             system_program, system_transaction,
-            timing::timestamp,
         },
-        solana_streamer::socket::SocketAddrSpace,
-        std::sync::{atomic::AtomicBool, mpsc::channel},
+        std::sync::mpsc::channel,
     };
 
     #[test]
@@ -341,8 +282,14 @@ mod test {
         let bank_forks = Arc::new(RwLock::new(BankForks::new(bank)));
         let (sender, receiver) = channel();
 
-        let send_tranaction_service =
-            SendTransactionService::new(tpu_address, &bank_forks, None, receiver, 1000, 1);
+        let send_tranaction_service = SendTransactionService::new::<NullTpuInfo>(
+            tpu_address,
+            &bank_forks,
+            None,
+            receiver,
+            1000,
+            1,
+        );
 
         drop(sender);
         send_tranaction_service.join().unwrap();
@@ -395,7 +342,7 @@ mod test {
                 None,
             ),
         );
-        let result = SendTransactionService::process_transactions(
+        let result = SendTransactionService::process_transactions::<NullTpuInfo>(
             &working_bank,
             &root_bank,
             &send_socket,
@@ -418,7 +365,7 @@ mod test {
             rooted_signature,
             TransactionInfo::new(rooted_signature, vec![], working_bank.block_height(), None),
         );
-        let result = SendTransactionService::process_transactions(
+        let result = SendTransactionService::process_transactions::<NullTpuInfo>(
             &working_bank,
             &root_bank,
             &send_socket,
@@ -441,7 +388,7 @@ mod test {
             failed_signature,
             TransactionInfo::new(failed_signature, vec![], working_bank.block_height(), None),
         );
-        let result = SendTransactionService::process_transactions(
+        let result = SendTransactionService::process_transactions::<NullTpuInfo>(
             &working_bank,
             &root_bank,
             &send_socket,
@@ -469,7 +416,7 @@ mod test {
                 None,
             ),
         );
-        let result = SendTransactionService::process_transactions(
+        let result = SendTransactionService::process_transactions::<NullTpuInfo>(
             &working_bank,
             &root_bank,
             &send_socket,
@@ -498,7 +445,7 @@ mod test {
                 None,
             ),
         );
-        let result = SendTransactionService::process_transactions(
+        let result = SendTransactionService::process_transactions::<NullTpuInfo>(
             &working_bank,
             &root_bank,
             &send_socket,
@@ -577,7 +524,7 @@ mod test {
                 Some((nonce_address, durable_nonce)),
             ),
         );
-        let result = SendTransactionService::process_transactions(
+        let result = SendTransactionService::process_transactions::<NullTpuInfo>(
             &working_bank,
             &root_bank,
             &send_socket,
@@ -604,7 +551,7 @@ mod test {
                 Some((nonce_address, Hash::new_unique())),
             ),
         );
-        let result = SendTransactionService::process_transactions(
+        let result = SendTransactionService::process_transactions::<NullTpuInfo>(
             &working_bank,
             &root_bank,
             &send_socket,
@@ -633,7 +580,7 @@ mod test {
                 Some((nonce_address, Hash::new_unique())),
             ),
         );
-        let result = SendTransactionService::process_transactions(
+        let result = SendTransactionService::process_transactions::<NullTpuInfo>(
             &working_bank,
             &root_bank,
             &send_socket,
@@ -660,7 +607,7 @@ mod test {
                 Some((nonce_address, durable_nonce)),
             ),
         );
-        let result = SendTransactionService::process_transactions(
+        let result = SendTransactionService::process_transactions::<NullTpuInfo>(
             &working_bank,
             &root_bank,
             &send_socket,
@@ -688,7 +635,7 @@ mod test {
                 Some((nonce_address, Hash::new_unique())), // runtime should advance nonce on failed transactions
             ),
         );
-        let result = SendTransactionService::process_transactions(
+        let result = SendTransactionService::process_transactions::<NullTpuInfo>(
             &working_bank,
             &root_bank,
             &send_socket,
@@ -716,7 +663,7 @@ mod test {
                 Some((nonce_address, Hash::new_unique())), // runtime advances nonce when transaction lands
             ),
         );
-        let result = SendTransactionService::process_transactions(
+        let result = SendTransactionService::process_transactions::<NullTpuInfo>(
             &working_bank,
             &root_bank,
             &send_socket,
@@ -745,7 +692,7 @@ mod test {
                 Some((nonce_address, durable_nonce)),
             ),
         );
-        let result = SendTransactionService::process_transactions(
+        let result = SendTransactionService::process_transactions::<NullTpuInfo>(
             &working_bank,
             &root_bank,
             &send_socket,
@@ -773,7 +720,7 @@ mod test {
         let nonce_account =
             AccountSharedData::new_data(43, &new_nonce_state, &system_program::id()).unwrap();
         working_bank.store_account(&nonce_address, &nonce_account);
-        let result = SendTransactionService::process_transactions(
+        let result = SendTransactionService::process_transactions::<NullTpuInfo>(
             &working_bank,
             &root_bank,
             &send_socket,
@@ -790,116 +737,5 @@ mod test {
                 ..ProcessTransactionsResult::default()
             }
         );
-    }
-
-    #[test]
-    fn test_get_leader_tpus() {
-        let ledger_path = get_tmp_ledger_path!();
-        {
-            let blockstore = Blockstore::open(&ledger_path).unwrap();
-
-            let validator_vote_keypairs0 = ValidatorVoteKeypairs::new_rand();
-            let validator_vote_keypairs1 = ValidatorVoteKeypairs::new_rand();
-            let validator_vote_keypairs2 = ValidatorVoteKeypairs::new_rand();
-            let validator_keypairs = vec![
-                &validator_vote_keypairs0,
-                &validator_vote_keypairs1,
-                &validator_vote_keypairs2,
-            ];
-            let GenesisConfigInfo {
-                genesis_config,
-                mint_keypair: _,
-                voting_keypair: _,
-            } = create_genesis_config_with_vote_accounts(
-                1_000_000_000,
-                &validator_keypairs,
-                vec![10_000; 3],
-            );
-            let bank = Arc::new(Bank::new_for_tests(&genesis_config));
-
-            let (poh_recorder, _entry_receiver, _record_receiver) = PohRecorder::new(
-                0,
-                bank.last_blockhash(),
-                0,
-                Some((2, 2)),
-                bank.ticks_per_slot(),
-                &Pubkey::default(),
-                &Arc::new(blockstore),
-                &Arc::new(LeaderScheduleCache::new_from_bank(&bank)),
-                &Arc::new(PohConfig::default()),
-                Arc::new(AtomicBool::default()),
-            );
-
-            let node_keypair = Arc::new(Keypair::new());
-            let cluster_info = Arc::new(ClusterInfo::new(
-                ContactInfo::new_localhost(&node_keypair.pubkey(), timestamp()),
-                node_keypair,
-                SocketAddrSpace::Unspecified,
-            ));
-
-            let validator0_socket = SocketAddr::from(([127, 0, 0, 1], 1111));
-            let validator1_socket = SocketAddr::from(([127, 0, 0, 1], 2222));
-            let validator2_socket = SocketAddr::from(([127, 0, 0, 1], 3333));
-            let recent_peers: HashMap<_, _> = vec![
-                (
-                    validator_vote_keypairs0.node_keypair.pubkey(),
-                    validator0_socket,
-                ),
-                (
-                    validator_vote_keypairs1.node_keypair.pubkey(),
-                    validator1_socket,
-                ),
-                (
-                    validator_vote_keypairs2.node_keypair.pubkey(),
-                    validator2_socket,
-                ),
-            ]
-            .iter()
-            .cloned()
-            .collect();
-            let leader_info = LeaderInfo {
-                cluster_info,
-                poh_recorder: Arc::new(Mutex::new(poh_recorder)),
-                recent_peers: recent_peers.clone(),
-            };
-
-            let slot = bank.slot();
-            let first_leader =
-                solana_ledger::leader_schedule_utils::slot_leader_at(slot, &bank).unwrap();
-            assert_eq!(
-                leader_info.get_leader_tpus(1),
-                vec![recent_peers.get(&first_leader).unwrap()]
-            );
-
-            let second_leader = solana_ledger::leader_schedule_utils::slot_leader_at(
-                slot + NUM_CONSECUTIVE_LEADER_SLOTS,
-                &bank,
-            )
-            .unwrap();
-            let mut expected_leader_sockets = vec![
-                recent_peers.get(&first_leader).unwrap(),
-                recent_peers.get(&second_leader).unwrap(),
-            ];
-            expected_leader_sockets.dedup();
-            assert_eq!(leader_info.get_leader_tpus(2), expected_leader_sockets);
-
-            let third_leader = solana_ledger::leader_schedule_utils::slot_leader_at(
-                slot + (2 * NUM_CONSECUTIVE_LEADER_SLOTS),
-                &bank,
-            )
-            .unwrap();
-            let mut expected_leader_sockets = vec![
-                recent_peers.get(&first_leader).unwrap(),
-                recent_peers.get(&second_leader).unwrap(),
-                recent_peers.get(&third_leader).unwrap(),
-            ];
-            expected_leader_sockets.dedup();
-            assert_eq!(leader_info.get_leader_tpus(3), expected_leader_sockets);
-
-            for x in 4..8 {
-                assert!(leader_info.get_leader_tpus(x).len() <= recent_peers.len());
-            }
-        }
-        Blockstore::destroy(&ledger_path).unwrap();
     }
 }
